// web/platform-admin-web/src/api/generated/model/updateResourceRequestDtoScope.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * AI Base App Platform Service API
 * The AI Base App Platform Service API documentation
 * OpenAPI spec version: 1.0
 */

/**
 * 리소스 관리 범위
 */
export type UpdateResourceRequestDtoScope = typeof UpdateResourceRequestDtoScope[keyof typeof UpdateResourceRequestDtoScope];


export const UpdateResourceRequestDtoScope = {
  PLATFORM: 'PLATFORM',
  ORGANIZATION: 'ORGANIZATION',
} as const;